% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{as.character.rendered_plot}
\alias{as.character.rendered_plot}
\title{Convert a rendered_plot object to a character}
\usage{
\method{as.character}{rendered_plot}(x, ...)
}
\arguments{
\item{x}{the rendered_plot}

\item{...}{not used}
}
\value{
a named vector
}
\description{
Convert a rendered_plot object to a character
}
