% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-utils.R
\name{as.character.rendered_table}
\alias{as.character.rendered_table}
\title{Convert a rendered_table object to a character}
\usage{
\method{as.character}{rendered_table}(x, ...)
}
\arguments{
\item{x}{the rendered_table}

\item{...}{not used}
}
\value{
a named vector
}
\description{
Convert a rendered_table object to a character
}
