% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-struct-list.R
\name{as.struct_list}
\alias{as.struct_list}
\title{Cast to a list of structures}
\usage{
as.struct_list(x, .class = NULL)
}
\arguments{
\item{x}{a list}

\item{.class}{the type of structures in the list}
}
\value{
a structured list
}
\description{
TODO: testing this
empty input (must specify .class)
struct_list - identity
simple struct - wrapped
list of struct_lists is flattened
list of structs is wrapped
nested plain lists / struct_lists are flattened
}
\concept{structures}
