% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{bind_rows_with_factors}
\alias{bind_rows_with_factors}
\title{Bind rows for colums with factors}
\usage{
bind_rows_with_factors(...)
}
\arguments{
\item{...}{a list of dataframes}
}
\value{
the union of those dataframes. Factor levels are combined with a superset of all levels
}
\description{
Bind_rows works until there are factors with a set of different levels then it throws a
wobbly. This handles that particular situation by combining factor levels.
}
\examples{
library(tidyverse)
bind_rows_with_factors(iris,
 ggplot2::diamonds \%>\% dplyr::rename(Species = cut)) \%>\%
 dplyr::pull(Species) \%>\%
 levels()
}
