% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_delete_stale}
\alias{cache_delete_stale}
\title{Delete stale files in a cache}
\usage{
cache_delete_stale(
  .cache = getOption("ggrrr.cache.dir", default = rappdirs::user_cache_dir("ggrrr")),
  .prefix = ".*",
  .stale = getOption("ggrrr.cache.stale", default = Inf)
)
}
\arguments{
\item{.cache}{the location of the cache as a directory. May get its value from options("ggrrr.cache.dir") or the default value of rappdirs::user_cache_dir("ggrrr")}

\item{.prefix}{a name of the operation so that you can namespace the cached files and do selective clean up operations on them}

\item{.stale}{the length of time in days to keep cached data before considering it as stale.}
}
\value{

}
\description{
Delete stale files in a cache
}
