% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_download}
\alias{cache_download}
\title{Download a file into a local cache.}
\usage{
cache_download(
  url,
  ...,
  .nocache = getOption("ggrrr.disable.cache", default = FALSE),
  .cache = getOption("ggrrr.download.dir", default =
    rappdirs::user_cache_dir("ggrrr-download")),
  .stale = getOption("ggrrr.cache.stale", default = Inf)
)
}
\arguments{
\item{url}{the url to download}

\item{...}{ignored}

\item{.nocache}{if set to TRUE all caching is disabled}

\item{.cache}{the location of the downloaded files}

\item{.stale}{how long to leave this file before replacing it.}
}
\value{
the path to the downloaded file
}
\description{
Download a file into a local cache.
}
