% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cached}
\alias{cached}
\title{A simple passthrough cache for complex or long running operations}
\usage{
cached(
  .expr = { },
  ...,
  .nocache = getOption("ggrrr.disable.cache", default = FALSE),
  .cache = getOption("ggrrr.cache.dir", default = rappdirs::user_cache_dir("ggrrr")),
  .prefix = getOption("ggrrr.item.prefix", default = "cached"),
  .stale = getOption("ggrrr.cache.stale", default = Inf)
)
}
\arguments{
\item{.expr}{the code the output of which requires caching. Other than a return value this should not create side effects or change global variables.}

\item{...}{inputs that the code in expr depends on and changes in which require the code re-running, Could be Sys.Date()}

\item{.nocache}{an option to defeat the caching which can be set globally as options("ggrrr.disable.cache"=TRUE)}

\item{.cache}{the location of the cache as a directory. May get its value from options("ggrrr.cache.dir") or the default value of rappdirs::user_cache_dir("ggrrr")}

\item{.prefix}{a name of the operation so that you can namespace the cached files and do selective clean up operations on them}

\item{.stale}{the length of time in days to keep cached data before considering it as stale. can also be set by options("ggrrr.cache.stale")}
}
\value{
the output of .expr which will usually be a value
}
\description{
executes expr and saves the output as an RDS file indexed by code in expr and the hash variable (which should contain any varying inputs)
}
\examples{
colName = "Petal.Width"
{
  iris[[colName]]
} \%>\% ggrrr::cached(iris, colName, .prefix="example", .cache=tempdir())
}
