% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utils.R
\name{code_snip}
\alias{code_snip}
\title{Display a code snippet}
\usage{
code_snip(
  type,
  filename,
  startMatches = "START",
  endMatches = "END",
  includeStart = FALSE,
  includeEnd = FALSE,
  sep = "\\n...\\n"
)
}
\arguments{
\item{type}{the code type}

\item{filename}{the source code file}

\item{startMatches}{a regex that matched start lines}

\item{endMatches}{a regex that matches end lines}

\item{includeStart}{is the regex inclusive of the start line or not}

\item{includeEnd}{is the regex inclusive of the end line or not}

\item{sep}{a seperator}
}
\value{
a text string of the selected code
}
\description{
Pulls out a code snippet based on a start and end
tags as comments within the code
}
\concept{output}
