% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utils.R
\name{code_snip_by_line}
\alias{code_snip_by_line}
\title{Display a code snippet}
\usage{
code_snip_by_line(type, filename, starts = 1, ends = Inf, sep = "\\n...\\n")
}
\arguments{
\item{type}{The code type (as understood by the minted latex plugin)}

\item{filename}{The source code filename}

\item{starts}{a vector of start indices ( as line numbers )}

\item{ends}{a vector of end indices ( as line numbers )}

\item{sep}{a separator}
}
\value{
a formatted string based on the file
}
\description{
Pulls out a code snippet based on a vector of start
and end lines.
}
\concept{output}
