% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-deprecated-dependency-utils.R
\name{cran}
\alias{cran}
\title{Make sure packages available on CRAN are installed}
\usage{
cran(cran_deps)
}
\arguments{
\item{cran_deps}{a vector of package names}
}
\value{
nothing
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# cran("tidyverse")
}
