% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-file-output-utils.R
\name{data_supplement}
\alias{data_supplement}
\title{Create a function list that allows for supplementary tables (as huxtables) to be added to a XLSX output file.}
\usage{
data_supplement(
  ...,
  filename = "supplementary-material.xlsx",
  out = ggrrr::outputter(...),
  nameGlue = "Supplementary Table {index}"
)
}
\arguments{
\item{...}{output location options will be passed to outputter(...) to define the location of the file}

\item{filename}{the xlsx filename}

\item{out}{an outputter (defaults to a default outputter )}

\item{nameGlue}{What will the tables be named}
}
\value{
a list of 2 functions.
$add_table(hux, caption, footnote, index), which takes a huxtable, caption, and index a writes the huxtable into a supplementary.
$write() which writes the collection of tables to the excel file.
}
\description{
This function encapsulates a excel output file as a destination for data tables. With the output of this
function you can add extra data to the supplement as a new sheet, or you can write the spreadsheet to disk.
When each data table is written either the table can be written silently or returned so that it is included
in a knitr document. This is controlled by `option("hide.supplementary.tables"=TRUE)`.
}
