% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{dbeta2}
\alias{dbeta2}
\title{The Beta Distribution}
\usage{
dbeta2(x, prob, kappa, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{prob}{the mean probability (vectorised)}

\item{kappa}{a coefficient of variation. where 0 is no variability and 1 is
maximally variability (vectorised)}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
\code{dbeta} gives the density, \code{pbeta} the distribution
  function, \code{qbeta} the quantile function, and \code{rbeta}
  generates random deviates.

  Invalid arguments will result in return value \code{NaN}, with a warning.

  The length of the result is determined by \code{n} for
  \code{rbeta}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.
}
\description{
Density, distribution function, quantile function and random
  generation for the Beta distribution with parameters \code{shape1} and
  \code{shape2} (and optional non-centrality parameter \code{ncp}).
}
\examples{
dbeta2(c(0.25,0.5,0.75), 0.5, 0.25)
}
\seealso{
\code{\link[stats:Beta]{stats::dbeta()}}
}
\concept{distributions}
