% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{dnbinom2}
\alias{dnbinom2}
\title{The Negative Binomial Distribution}
\source{
\code{dnbinom} computes via binomial probabilities, using code
  contributed by Catherine Loader (see \code{\link[stats]{dbinom}}).

  \code{pnbinom} uses \code{\link[stats]{pbeta}}.

  \code{qnbinom} uses the Cornish--Fisher Expansion to include a skewness
  correction to a normal approximation, followed by a search.

  \code{rnbinom} uses the derivation as a gamma mixture of Poisson
  distributions, see

  Devroye, L. (1986) \emph{Non-Uniform Random Variate Generation.}
  Springer-Verlag, New York. Page 480.
}
\usage{
dnbinom2(x, mean, sd = sqrt(mean), log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{mean}{the mean value on the true scale (vectorised)}

\item{sd}{the standard deviation on the true scale (vectorised)}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
\code{dnbinom} gives the density,
  \code{pnbinom} gives the distribution function,
  \code{qnbinom} gives the quantile function, and
  \code{rnbinom} generates random deviates.

  Invalid \code{size} or \code{prob} will result in return value
  \code{NaN}, with a warning.

  The length of the result is determined by \code{n} for
  \code{rnbinom}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.

  \code{rnbinom} returns a vector of type \link{integer} unless generated
  values exceed the maximum representable integer when \code{\link{double}}
  values are returned.
}
\description{
Density, distribution function, quantile function and random
  generation for the negative binomial distribution with parameters
  \code{size} and \code{prob}.
}
\details{
The negative binomial distribution with \code{size} \eqn{= n} and
  \code{prob} \eqn{= p} has density
  \deqn{
    p(x) = \frac{\Gamma(x+n)}{\Gamma(n) x!} p^n (1-p)^x}{
    \\Gamma(x+n)/(\\Gamma(n) x!) p^n (1-p)^x}
  for \eqn{x = 0, 1, 2, \ldots}, \eqn{n > 0} and \eqn{0 < p \le 1}.

  This represents the number of failures which occur in a sequence of
  Bernoulli trials before a target number of successes is reached.
  The mean is \eqn{\mu = n(1-p)/p} and variance \eqn{n(1-p)/p^2}.

  A negative binomial distribution can also arise as a mixture of
  Poisson distributions with mean distributed as a gamma distribution
  (see \code{\link[stats]{pgamma}}) with scale parameter \code{(1 - prob)/prob}
  and shape parameter \code{size}.  (This definition allows non-integer
  values of \code{size}.)

  An alternative parametrization (often used in ecology) is by the
  \emph{mean} \code{mu} (see above), and \code{size}, the \emph{dispersion
  parameter}, where \code{prob} = \code{size/(size+mu)}.  The variance
  is \code{mu + mu^2/size} in this parametrization.

  If an element of \code{x} is not integer, the result of \code{dnbinom}
  is zero, with a warning.

  The case \code{size == 0} is the distribution concentrated at zero.
  This is the limiting distribution for \code{size} approaching zero,
  even if \code{mu} rather than \code{prob} is held constant.  Notice
  though, that the mean of the limit distribution is 0, whatever the
  value of \code{mu}.

  The quantile is defined as the smallest value \eqn{x} such that
  \eqn{F(x) \ge p}, where \eqn{F} is the distribution function.
}
\examples{
dnbinom2(0:5, 2, sqrt(2))
}
\seealso{
\link[stats]{Distributions} for standard distributions, including
  \code{\link[stats]{dbinom}} for the binomial, \code{\link[stats]{dpois}} for the
  Poisson and \code{\link[stats]{dgeom}} for the geometric distribution, which
  is a special case of the negative binomial.
}
\concept{distributions}
