% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{.embed_fonts}
\alias{.embed_fonts}
\title{Embed font families into pdf}
\usage{
.embed_fonts(pdfFile, fonts)
}
\arguments{
\item{pdfFile}{the pdf}

\item{fonts}{the font families}
}
\value{
nothing
}
\description{
Embeds a set of font files relating to the specified families into the
pdf file, from local file storage.
}
\keyword{internal}
