% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.find_chrome}
\alias{.find_chrome}
\title{Find Google Chrome or Chromium in the system}
\usage{
.find_chrome()
}
\value{
A character string with the path of chrome or an error.
}
\description{
On Windows, this function tries to find Chrome from the registry. On macOS,
it returns a hard-coded path of Chrome under \file{/Applications}. On Linux,
it searches for \command{chromium-browser} and \command{google-chrome} from
the system's \var{PATH} variable.
}
\details{
This function is borrowed from `pagedown` here to avoid the dependency:
https://github.com/rstudio/pagedown/blob/main/R/chrome.R which is MIT
licensed. We note that there is no check made on MacOS that the path is
correct, and since it may be installed elsewhere this is a potential point
of failure. Must assume that calls to chrome may fail.
}
\concept{output}
\keyword{internal}
