% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-stack-tools.R
\name{.find_namespace}
\alias{.find_namespace}
\title{Find the package environment for a given environment or function}
\usage{
.find_namespace(env = rlang::caller_env())
}
\arguments{
\item{env}{an environment or function.}
}
\value{
a namespaced package environment
}
\description{
This searches the environment stack looking for the first environment
that defines a package. This is the package that the function is defined in.
If no namespace environment is found then the function is in the global
environment.
}
\keyword{internal}
