% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-stack-tools.R
\name{.get_fn_name}
\alias{.get_fn_name}
\title{Get the defined name of a function}
\usage{
.get_fn_name(fn)
}
\arguments{
\item{fn}{a function}
}
\value{
the name as a string, as the function was defined
}
\description{
Get the defined name of a function
}
\keyword{internal}
