% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggplot-utils.R
\name{.gg_breaks_log1p}
\alias{.gg_breaks_log1p}
\title{A scales breaks generator for log1p scales}
\usage{
.gg_breaks_log1p(n = 5, base = 10)
}
\arguments{
\item{n}{the number of breaks}

\item{base}{the base for the breaks}
}
\value{
a function for ggplot scale breaks
}
\description{
A scales breaks generator for log1p scales
}
\concept{ggplot}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
ggplot2::ggplot(ggplot2::diamonds, ggplot2::aes(x=price))+
  ggplot2::geom_density()+
  ggplot2::scale_x_continuous(trans="log1p", breaks=.gg_breaks_log1p())
}\if{html}{\out{</div>}}
}

