% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggplot-utils.R
\name{.gg_fonts_missing}
\alias{.gg_fonts_missing}
\title{Check if any fonts listed are missing}
\usage{
.gg_fonts_missing(family)
}
\arguments{
\item{family}{the font family}
}
\value{
`TRUE` if missing fonts detected
}
\description{
Check if any fonts listed are missing
}
\concept{ggplot}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
.gg_fonts_missing("Arial")
.gg_fonts_missing(c("Roboto","Kings","ASDASDAS"))
}\if{html}{\out{</div>}}
}

