% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggplot-utils.R
\name{.gg_scale_x_logit}
\alias{.gg_scale_x_logit}
\title{A logit x scale}
\usage{
.gg_scale_x_logit(..., n = 5, sf = 2)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}

\item{n}{the number of major breaks}

\item{sf}{significant figures}
}
\value{
a ggplot scale
}
\description{
A logit x scale
}
\concept{ggplot}
\keyword{internal}
