% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggplot-utils.R
\name{.gg_scale_x_percent}
\alias{.gg_scale_x_percent}
\title{A percentage x scale}
\usage{
.gg_scale_x_percent(..., sf = 2)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}

\item{sf}{significant figures}
}
\value{
a ggplot scale
}
\description{
A percentage x scale
}
\concept{ggplot}
\keyword{internal}
