% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggplot-utils.R
\name{.gg_set_size_defaults}
\alias{.gg_set_size_defaults}
\title{Set sizes in ggplot uniformly}
\usage{
.gg_set_size_defaults(
  lineSize = 0.5,
  fontSizePts = 4 + lineSize * 8,
  font = "Roboto",
  size = lineSize * 2
)
}
\arguments{
\item{lineSize}{the width of lines}

\item{fontSizePts}{the size of labels and other on plot text in pts.}

\item{font}{the font family name}

\item{size}{the size of points (the default size aesthetic)}
}
\value{
nothing
}
\description{
Set the default sizes of lines, points and fonts in ggplot geoms, and text labels in ggplot axes
to get a single consistent look and feel.
}
\concept{ggplot}
\keyword{internal}
