% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggplot-utils.R
\name{.gg_substitute_fonts}
\alias{.gg_substitute_fonts}
\title{Pick a locally installed font family that matches requested}
\usage{
.gg_substitute_fonts(family)
}
\arguments{
\item{family}{the font family requested}
}
\value{
a mapping as a named list of font families that are present on the
  system (names are the requested font family)
}
\description{
Pick a locally installed font family that matches requested
}
\examples{
try({
.gg_substitute_fonts(c("Roboto","Arial","Kings","Unmatched"))
})
}
\concept{ggplot}
\keyword{internal}
