% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggplot-utils.R
\name{.gg_transform_logit}
\alias{.gg_transform_logit}
\title{logit scale}
\usage{
.gg_transform_logit(n = 5, ...)
}
\arguments{
\item{n}{number of breas}

\item{...}{not used}
}
\value{
A scales object
}
\description{
Perform logit scaling with right axis formatting. To not be
used directly but with ggplot (e.g. ggplot2::scale_y_continuous(trans =
"logit")
}
\concept{ggplot}
\keyword{internal}
