% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.here}
\alias{.here}
\title{Drop in replacement for \code{here} (\code{here} pkg)}
\usage{
.here(..., projRoot = .locate_project())
}
\arguments{
\item{...}{the relative path within the project}

\item{projRoot}{the project root - defaults to \code{.locate_project()}}
}
\value{
a path
}
\description{
Drop in replacement for \code{here} (\code{here} pkg)
}
\concept{output}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
.here("vignettes")
}\if{html}{\out{</div>}}
}

