% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-huxtable-utils.R
\name{.hux_default_layout}
\alias{.hux_default_layout}
\title{A tidy article theme for huxtables}
\usage{
.hux_default_layout(
  hux,
  defaultFontSize = 8,
  defaultFont = "Roboto",
  headerRows = 1
)
}
\arguments{
\item{hux}{a huxtable object}

\item{defaultFontSize}{default size of font in points (8)}

\item{defaultFont}{the font family name}

\item{headerRows}{the number of rows that are headers}
}
\value{
the formatted huxtable.
}
\description{
The main aim is to get something that works with google docs when you copy and paste.
}
\examples{
try({
hux = iris \%>\% .hux_default_layout()
})
}
\concept{huxtable}
\keyword{internal}
