% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-text-formatting.R
\name{.if_na}
\alias{.if_na}
\title{Format a value replacing NA values}
\usage{
.if_na(x, na.replace = "―", ...)
}
\arguments{
\item{x}{a vector of values}

\item{na.replace}{a value to replace NA values with}

\item{...}{
  Arguments passed on to \code{\link[base:format]{base::format}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
a vector of character values
}
\description{
Format a value replacing NA values
}
\examples{
try({
.if_na( c(1,2,NA,4,5)/3, digits=2 )
})
}
\keyword{internal}
