% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-text-formatting.R
\name{.if_present}
\alias{.if_present}
\title{Format a value omitting NA values}
\usage{
.if_present(x, ...)
}
\arguments{
\item{x}{a vector of values}

\item{...}{
  Arguments passed on to \code{\link[base:format]{base::format}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
a vector of character values
}
\description{
Format a value omitting NA values
}
\examples{
try({
.if_present( c(1,2,NA,4,5)/3, digits=2 )
})
}
\keyword{internal}
