% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.is_document_output}
\alias{.is_document_output}
\title{Check knitr output is a word type document}
\usage{
.is_document_output()
}
\value{
true if the knitr output target is doc or odt
}
\description{
Check knitr output is a word type document
}
\concept{output}
\keyword{internal}
