% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.is_knitting}
\alias{.is_knitting}
\title{Check if the current code is being executed during document knitting}
\usage{
.is_knitting()
}
\value{
TRUE if the whole document is being knitted.
FALSE if running in chunk in RStudio, or otherwise outside of a knitr
}
\description{
Check if the current code is being executed during document knitting
}
\concept{output}
\keyword{internal}
