% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.is_output}
\alias{.is_output}
\title{Check knitr output is of a specified output type}
\usage{
.is_output(outputFormats)
}
\arguments{
\item{outputFormats}{the name of a rmarkdown output type. This is the
value given in the `output:` part of the yaml metadata.}
}
\value{
true or false
}
\description{
Check knitr output is of a specified output type
}
\concept{output}
\keyword{internal}
