% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.is_running_in_chunk}
\alias{.is_running_in_chunk}
\title{Check if the current code is being executed in a document chunk in
an RStudio session or when being knitted.}
\usage{
.is_running_in_chunk()
}
\value{
TRUE is being knitted OR running in chunk in RStudio
FALSE if not interactive or interactive but in console in RStudio
}
\description{
Check if the current code is being executed in a document chunk in
an RStudio session or when being knitted.
}
\concept{output}
\keyword{internal}
