% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.is_running_in_console}
\alias{.is_running_in_console}
\title{Check if the current code is being executed in rstudio console or from
a console outside rstudio}
\usage{
.is_running_in_console()
}
\value{
TRUE is being run in console
FALSE if being knitted, or being run
}
\description{
Check if the current code is being executed in rstudio console or from
a console outside rstudio
}
\concept{output}
\keyword{internal}
