% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.is_unknown_output}
\alias{.is_unknown_output}
\title{Check knitr output is not a known output forma}
\usage{
.is_unknown_output()
}
\value{
true or false
}
\description{
Check knitr output is not a known output forma
}
\concept{output}
\keyword{internal}
