% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.is_viewer_available}
\alias{.is_viewer_available}
\title{Check if a viewer is available}
\usage{
.is_viewer_available()
}
\value{
true or false
}
\description{
Check if a viewer is available
}
\concept{output}
\keyword{internal}
