% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-stack-tools.R
\name{.lax_as_function}
\alias{.lax_as_function}
\title{Permissive as_function replacement.}
\usage{
.lax_as_function(fn)
}
\arguments{
\item{fn}{something to interpret as a function or a primitive}
}
\value{
the function or primitive
}
\description{
like `rlang::as_function` but only interprets functions or formulae
and ignores primitives and characters.
}
\keyword{internal}
