% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.locate_project}
\alias{.locate_project}
\title{Find the root of the current project}
\usage{
.locate_project(inputFile = .this_script())
}
\arguments{
\item{inputFile}{a file to check the project root of.}
}
\value{
the file path of the root of the project
}
\description{
gets a reverse directory listing and then finds the lowest directory
that contains an indication of there being a project in the directory
this should work where the `here` package does not
(particularly when knitting in CRAN checks for example, or running in a
non package project)
}
\examples{
try({
.locate_project()
})
}
\concept{output}
\keyword{internal}
