% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-memoise.R
\name{.memoise}
\alias{.memoise}
\title{Memoise a function with an in memory cache}
\usage{
.memoise(fn, .cache_errors = FALSE, .debug = FALSE)
}
\arguments{
\item{fn}{a self contained function. This should not reference anything
outside of its own parameters, and should be deterministic.}
}
\value{
a function that takes the
}
\description{
Options to control:
- options(".memoise.disabled"): no new functions are memoised. existing
  memoised functions will recalculate the values
- options(".memoise.recalculate"):
}
\keyword{internal}
