% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.optional_fn}
\alias{.optional_fn}
\title{Get an optional function without triggering a CRAN warning}
\usage{
.optional_fn(
  pkg,
  name,
  alt = function(...) {
     stop("function `", pkg, "::", name, "(...)` not available")

    }
)
}
\arguments{
\item{pkg}{the package name}

\item{name}{the function you wish to use}

\item{alt}{a function that can be used instead}
}
\value{
the function you want if available or the alternative
}
\description{
You want to use a function if it is installed but don't want it to
be installed as part of your package and you don't want to reference it
as part of the Imports or Suggests fields in a package DESCRIPTION.
}
\examples{
fn = .optional_fn("openssl", "md5", digest::digest)
as.character(fn(as.raw(c(1,2,3))))
}
\keyword{internal}
