% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-text-formatting.R
\name{.pdf_safe}
\alias{.pdf_safe}
\title{Switch UTF-8 into plain text when using the pdf device}
\usage{
.pdf_safe(label, alt = label)
}
\arguments{
\item{label}{a UTF8 label}

\item{alt}{alternative for the pdf device}
}
\value{
the same string or a non UTF alternative if currently using the
  legacy pdf device
}
\description{
The plain `grDevices::pdf()` device is the default when running examples via CRAN R CMD
check. It throws warnings and sometimes errors that other devices do not when
encountering UTF-8 characters in plot objects (due to a warning from `grid`).
The resulting behaviour is R version dependent. It is basically impossible to
get round these in your function examples and you either decide not to run
them or only run them with non UTF-8 content. This will make that decision at
runtime and provide a transliterated alternative for the `pdf` device in the
context of a function example in a CRAN check.
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
.pdf_safe("test")
.pdf_safe("\u00B1\u221E")
ggplot2::ggplot()+ggplot2::xlab(.pdf_safe("\u00B1\u221E"))
}\if{html}{\out{</div>}}
}

