% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.pmixbeta}
\alias{.pmixbeta}
\title{The cumulative density function of a mixture of beta distributions}
\usage{
.pmixbeta(q, alphas, betas, weights = 1, na.rm = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{alphas}{a vector of beta distribution alphas}

\item{betas}{a vector of gamma distribution betas}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions which have NA for shape or rate}
}
\value{
the pdf of the mixture distribution.
}
\description{
The cumulative density function of a mixture of beta distributions
}
\concept{distributions}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
.pmixbeta(q=c(2,20), alphas=c(10,13,14), betas=c(1,1,1), weights=c(2,2,3))
}\if{html}{\out{</div>}}
}

