% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.qmixbeta}
\alias{.qmixbeta}
\title{A quantile function for a mixture of gamma distributions}
\usage{
.qmixbeta(
  p,
  alphas,
  betas,
  weights = rep(1, length(alphas)),
  na.rm = FALSE,
  approx = TRUE,
  ...
)
}
\arguments{
\item{p}{vector of probabilities.}

\item{alphas}{a vector of gamma distribution shapes}

\item{betas}{a vector of gamma distribution rates}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions with NA values for mean or sd}

\item{approx}{use Cornish Fisher moment based approximation if available}

\item{...}{passed on to internal function, `samples=TRUE` to force random sampling}
}
\value{
the value of the `p`th quantile
}
\description{
A quantile function for a mixture of gamma distributions
}
\examples{
try({
  .qmixbeta(p=c(0.025,0.5,0.975), alphas=c(10,13,14), betas=c(1,1,2))
})
}
\concept{distributions}
\keyword{internal}
