% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.qmixlnorm}
\alias{.qmixlnorm}
\title{A quantile function for a mixture of log normal distributions}
\usage{
.qmixlnorm(
  p,
  meanlogs,
  sdlogs,
  weights = rep(1, length(meanlogs)),
  na.rm = FALSE,
  method = c("exact", "samples", "moments"),
  ...
)
}
\arguments{
\item{p}{vector of probabilities.}

\item{meanlogs}{a vector of log normal distribution means}

\item{sdlogs}{a vector of log normal distribution sds}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions with NA values for mean or sd}

\item{method}{one of \code{exact} (solve with uniroot), \code{samples} (random resampling), \code{moments} (Cornish Fisher approximation)}

\item{...}{passed on to internal function, \code{seed=XXX} to fix random seed}
}
\value{
the value of the \code{p}th quantile
}
\description{
A quantile function for a mixture of log normal distributions
}
\concept{distributions}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
.qmixlnorm(p=c(0.025,0.5,0.975), meanlogs=c(1,1.3,1.4), sdlogs=c(0.1,0.1,0.2))
.qmixlnorm(p=c(0.025,0.5,0.975), meanlogs=c(1,1.3,1.4), sdlogs=c(0.1,0.1,0.2), method="samples")
.qmixlnorm(p=c(0.025,0.5,0.975), meanlogs=c(1,1.3,1.4), sdlogs=c(0.1,0.1,0.2), method="moments")

}\if{html}{\out{</div>}}
}

