% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-text-formatting.R
\name{.sprintf_data}
\alias{.sprintf_data}
\title{`sprintf` with a list(ish) input}
\usage{
.sprintf_data(fmt, params, ..., na.replace = "―")
}
\arguments{
\item{fmt}{a sprintf format string}

\item{params}{the inputs as a list or a dataframe (rather than an enumeration of individual numbers)}

\item{...}{
  Named arguments passed on to \code{\link[=.sprintf_dp]{.sprintf_dp}}\describe{
    \item{\code{sep}}{a character to use a decimal point.}
}}

\item{na.replace}{a value to replace NA values with.}
}
\value{
the formatted string
}
\description{
A variant of sprintf that work well with inputs that are in the format of a list
or a dataframe. Examples of which are the quantile functions. This falls
in between `sprintf` and `glue::glue_data` in complexity.
}
\keyword{internal}
