% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-text-formatting.R
\name{.sprintf_dp}
\alias{.sprintf_dp}
\title{Replaces decimal points in sprintf output.}
\usage{
.sprintf_dp(fmt, ..., sep = getOption("OutDec", "."))
}
\arguments{
\item{fmt}{a sprintf format string}

\item{...}{
  Named arguments passed on to \code{\link[base:sprintf]{base::sprintf}}\describe{
    \item{\code{...}}{values to be passed into \code{fmt}.  Only logical,
    integer, real and character vectors are supported, but some coercion
    will be done: see the \sQuote{Details} section. Up to 100.}
\item{\code{trim,domain}}{see \code{\link[base]{gettext}}.}
}}

\item{sep}{a character to use a decimal point.}
}
\value{
the formatted string with decimal points as specified
}
\description{
`sprintf` will respect decimal point conventions only if you switch locale.
This is unwieldy and not in line with `format`. This wrapper around `sprintf`
replaces decimal points with a `sep` character of your choice, sensible choices
for which are `,` or `UTF8+00B7`. It will not try and replace things that are
not syntactially numbers.
}
\examples{
try({
.sprintf_dp("\%1.2f",1:3/3, sep="\u00B7")
.sprintf_dp("\%1.2f-\%1.2f", 1:3/3, 1:3, sep="\u00B7")
.sprintf_dp("\%s \%1.2f-\%1.2f", "A.1.2", 1:3/3, 1:3, sep="\u00B7")
})
}
\keyword{internal}
