% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{.substitute_fonts}
\alias{.substitute_fonts}
\title{Pick a locally installed font family that matches requested}
\usage{
.substitute_fonts(family, quiet = TRUE)
}
\arguments{
\item{family}{the font family requested}

\item{quiet}{do not print warnings.}
}
\description{
Pick a locally installed font family that matches requested
}
\examples{
try({
.substitute_fonts(c("Roboto","Arial","Kings","Unmatched"))
})
}
