% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-memoise.R
\name{.unmemoise}
\alias{.unmemoise}
\title{Unmemoise a function and drop its cache.}
\usage{
.unmemoise(fn)
}
\arguments{
\item{fn}{the memoised or unmemoised function.}
}
\value{
the unmemoised function if it was found. If the function was not
found then the input is returned with a message.
}
\description{
Unmemoise a function and drop its cache.
}
\keyword{internal}
