% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{drawDetails.watermark}
\alias{drawDetails.watermark}
\title{Internal function for drawing watermark on ggplots}
\usage{
drawDetails.watermark(x, rot = 45, ...)
}
\arguments{
\item{x}{the grob}

\item{rot}{degrees of rotation}

\item{...}{ignored}
}
\value{
a grid object
}
\description{
Internal function for drawing watermark on ggplots
}
