% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{dwedge}
\alias{dwedge}
\title{Wedge distribution}
\usage{
dwedge(x, a, lower.tail = TRUE, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{a}{a gradient from -2 (left skewed) to 2 (right skewed)}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}
otherwise \code{P[X>x]}.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The wedge distribution has a domain of 0 to 1 and has a linear probability
density function over that domain.
}
\details{
The \code{rwedge} can be combined with quantile functions to
skew standard distributions, or introduce correlation or down weight
certain parts of the distribution.
}
\examples{

pwedge(seq(0,1,0.1), a=1)
dwedge(seq(0,1,0.1), a=1)
qwedge(c(0.25,0.5,0.75), a=-1)

stats::cor(
  stats::qnorm(rwedge(1000, a=2)),
  stats::qnorm(rwedge(1000, a=-2))
)
}
\concept{distributions}
