% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{full_seq_dates}
\alias{full_seq_dates}
\title{Expand a data vector to the full range}
\usage{
full_seq_dates(dates)
}
\arguments{
\item{dates}{a vector of dates, possibly including NA values}
}
\value{
a vector of dates for every day between the minimum and maximum of dates
}
\description{
Convert a vector of observation dates to a ordered sequence of every day in the time series
}
\examples{
full_seq_dates(c("2020-01-01","2020-02-01","2020-01-15","2020-02-01",NA))
}
