% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{get_value_sets}
\alias{get_value_sets}
\title{This function examines a dataframe and returns a list of the columns with sub-lists as all the options for factors}
\usage{
get_value_sets(df)
}
\arguments{
\item{df}{a dataframe to examine}
}
\value{
a list of lists with the column name and the factor levels as list
}
\description{
This function examines a dataframe and returns a list of the columns with sub-lists as all the options for factors
}
