% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-file-output-utils.R
\name{gg_find_webfonts}
\alias{gg_find_webfonts}
\title{Find webfonts for a ggplot}
\usage{
gg_find_webfonts(plot)
}
\arguments{
\item{plot}{a ggplot}
}
\value{
a list of css webfont specifications
}
\description{
Resolves any missing fonts in a ggplot and tries to resolve them
against webfont providers (Brick and Google Fonts) locally cacheing
them. This function is the default for `gg_save_as`
}
