% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{gg_formatted_table}
\alias{gg_formatted_table}
\title{Display a long format table as a ggplot object. This is useful if you want to combine a formatted table with a plot in a}
\usage{
gg_formatted_table(
  longFormatTable,
  colWidths = NULL,
  tableWidthInches = 5.9,
  ...
)
}
\arguments{
\item{longFormatTable}{a table - usually converted using as.long_format_table()}

\item{colWidths}{(optional) the relative widths of the columns.}

\item{tableWidthInches}{the maximum desired width of the plot. Text will be scaled to fit this width.}

\item{...}{passed to as.long_format_table if and only if the input is not already in that format.}
}
\value{
a ggplot object
}
\description{
Display a long format table as a ggplot object. This is useful if you want to combine a formatted table with a plot in a
}
