% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{gg_pedantic}
\alias{gg_pedantic}
\title{An opinionated set of defaults for plot styles with a focus on making plots compact, and minimally fussy, and ensuring fonts are consistent between axes and labels.}
\usage{
gg_pedantic(lineSize = 0.25, fontSize = 8, font = "Roboto")
}
\arguments{
\item{lineSize}{the default line and shape size in ggplot units}

\item{fontSize}{the base font size}

\item{font}{the default font name.}
}
\value{
nothing
}
\description{
An opinionated set of defaults for plot styles with a focus on making plots compact, and minimally fussy, and ensuring fonts are consistent between axes and labels.
}
