% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{gg_resize_legend}
\alias{gg_resize_legend}
\title{Make the legend smaller}
\usage{
gg_resize_legend(pointSize = 0.75, textSize = 6, spaceLegend = 0.75)
}
\arguments{
\item{pointSize}{- the ggplot size of lines or points}

\item{textSize}{- the size in pts of the text}

\item{spaceLegend}{- degree of spacing between items in the scale (defines overall size)}
}
\value{
a theme
}
\description{
Make the legend smaller
}
