% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{gg_save_as}
\alias{gg_save_as}
\title{Allows specific maximum dimensions with an optional target aspect ratio.
If no aspect ratio supplied then plot will be shaped to occupy whole area}
\usage{
gg_save_as(
  plot,
  filename = tempfile(),
  size = std_size$half,
  maxWidth = size$width,
  maxHeight = size$height,
  aspectRatio = maxWidth/maxHeight,
  formats = c("png", "pdf", "Rdata")
)
}
\arguments{
\item{plot}{a GGplot object or none}

\item{filename}{base of target filename (excuding extension).}

\item{maxWidth}{maximum width in inches}

\item{maxHeight}{maximum height in inches}

\item{aspectRatio}{defaults to maxWidth/maxHeight}
}
\description{
Allows specific maximum dimensions with an optional target aspect ratio.
If no aspect ratio supplied then plot will be shaped to occupy whole area
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveFigure(filename="the_filename",maxWidth=4,maxHeight=4)
}
\keyword{plot}
