% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{gg_set_size_defaults}
\alias{gg_set_size_defaults}
\title{Set the default sizes of lines and fonts in ggplot geoms}
\usage{
gg_set_size_defaults(
  lineSize = 0.5,
  fontSizePts = 4 + lineSize * 8,
  font = "Roboto"
)
}
\arguments{
\item{lineSize}{the width of lines and size of points (the default size aesthetic)}

\item{fontSizePts}{the size of labels and other on plot text in pts.}
}
\value{
nothing
}
\description{
Set the default sizes of lines and fonts in ggplot geoms
}
\examples{
gg_set_size_defaults(lineSize = 0.25)
}
