% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{gg_simple_table}
\alias{gg_simple_table}
\title{A simple table as a ggplot patchwork object, no customisation allowed}
\usage{
gg_simple_table(df, pts = 8, unwrapped = FALSE)
}
\arguments{
\item{df}{the dataframe with the table data}

\item{pts}{text size}

\item{unwrapped}{- set this to TRUE if you want to add to a patchwork and use wrap_plots(p,list(table))}
}
\description{
A simple table as a ggplot patchwork object, no customisation allowed
}
\examples{
...+simpleFigureTable(tibble(x=c(1,2,3),y=c(5,4,3)),pts=10)
}
\keyword{graph}
\keyword{layout}
