% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{gg_watermark}
\alias{gg_watermark}
\title{Add in a watermark to graphs}
\usage{
gg_watermark(
  lab = "DRAFT",
  disable = getOption("ggrrr.disable.watermark", default = FALSE)
)
}
\arguments{
\item{lab}{the watermark label (DRAFT)}

\item{disable}{- global option to disable all watermarks options("ggrrr.disable.watermark"=TRUE)}
}
\value{
a watermark layer
}
\description{
Add in a watermark to graphs
}
