% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrrr-package.R
\docType{package}
\name{ggrrr-package}
\alias{ggrrr}
\alias{ggrrr-package}
\title{ggrrr: Addressing Annoyances and Irritations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Visualisation hacks, tabular data helpers, fonts, caching, tidy data functions. It is an swiss army knife, jack of all trades.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://terminological.github.io/ggrrr/index.html}
  \item \url{https://github.com/terminological/ggrrr}
  \item \url{https://terminological.github.io/ggrrr/}
  \item Report bugs at \url{https://github.com/terminological/ggrrr/issues}
}

}
\author{
\strong{Maintainer}: Robert Challen \email{rob@terminolgical.co.uk} (\href{https://orcid.org/0000-0002-5504-7768}{ORCID})

}
\keyword{internal}
