% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{google_font}
\alias{google_font}
\title{Downloads caches and installs a font family from google}
\usage{
google_font(family, ...)
}
\arguments{
\item{family}{the google font family (e.g. Roboto) as found here: https://fonts.google.com/ (case sensitive)}

\item{...}{not used}
}
\value{
nothing
}
\description{
This checks if the file is already downloaded and if not pull the font file from google fonts.
If it does not exist it will throw an error. Once downloaded a font will be made available through
`showtext` and through `systemfonts` font packages.
}
