% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-file-output-utils.R
\name{html_pdf_converter}
\alias{html_pdf_converter}
\title{Convert html to pdf depending on what is available on the platform}
\usage{
html_pdf_converter(html, filename, maxWidth, maxHeight)
}
\arguments{
\item{html}{the html fragment}

\item{filename}{the pdf filename}

\item{maxWidth}{the maximum page width in inches}

\item{maxHeight}{the maximum page height in inches}
}
\value{
nothing called for side effects
}
\description{
If html2pdfr is installed it will use that by default, if not it will fall
back to a headless chrome instance.
}
