% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-huxtable-utils.R
\name{hux_auto_widths}
\alias{hux_auto_widths}
\title{Calculate a sensible column and table width for a huxtable based on its content.}
\usage{
hux_auto_widths(hux, target = "html", including_headers = FALSE)
}
\arguments{
\item{hux}{the huxtable}

\item{target}{the expected output (could be "docx"/"odt", "xlsx") which are the only options that matter}

\item{including_headers}{Should we try and fit the header contents as well (TRUE) or let those wrap (FALSE).}
}
\value{
the huxtable with the width options set.
}
\description{
Calculate a sensible column and table width for a huxtable based on its content.
}
\concept{huxtable}
