% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-utils.R
\name{hux_default_layout}
\alias{hux_default_layout}
\title{A tidy article theme for huxtables that works with google docs}
\usage{
hux_default_layout(
  hux,
  defaultFontSize = 8,
  defaultFont = "Roboto",
  headerRows = 1
)
}
\arguments{
\item{hux}{a huxtable object}

\item{defaultFontSize}{default size of font in points (8)}

\item{defaultFont}{the font name}
}
\description{
A tidy article theme for huxtables that works with google docs
}
\examples{
hux = huxtable(dataframe \%>\% select("col 1 title"=col1)) \%>\% defaultTableLayout()
}
\keyword{huxtable}
