% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-utils.R
\name{hux_insert_start}
\alias{hux_insert_start}
\title{Insert row at start maintaining format}
\usage{
hux_insert_start(hux, ..., fill = "", colspan = 1)
}
\arguments{
\item{hux}{a huxtable}

\item{...}{stuff to insert}

\item{fill}{padding}
}
\value{
a huxtable with row inserted at start in the same format
}
\description{
Insert row at start maintaining format
}
