% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-utils.R
\name{hux_nest_group}
\alias{hux_nest_group}
\title{Make a huxtable narrower}
\usage{
hux_nest_group(t, col = 1)
}
\arguments{
\item{t}{the huxtable}

\item{col}{the column index you want to nest into the row above}
}
\value{
a narrower huxtable
}
\description{
Make a huxtable narrower
}
