% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-file-output-utils.R
\name{hux_save_as}
\alias{hux_save_as}
\title{Save a table to a variety of formats}
\usage{
hux_save_as(...)
}
\arguments{
\item{...}{
  Named arguments passed on to \code{\link[=.hux_save_as]{.hux_save_as}}\describe{
    \item{\code{hux}}{the huxtable to save}
\item{\code{filename}}{the filename, which may omit the extension}
\item{\code{size}}{a `std_size` list entry}
\item{\code{maxWidth}}{or the maximum width in inches}
\item{\code{maxHeight}}{and either the maximum height in inches}
\item{\code{aspectRatio}}{or the minimum allowed aspect ratio}
\item{\code{formats}}{if the extension is omitted, all the formats described here will be saved. Currently supported outputs are "html","png","pdf","docx","xlsx"}
\item{\code{defaultFontSize}}{the default font size}
\item{\code{sheetname}}{if saving as an xlsx file.}
}}
}
\value{
the output depends on if the function is called in a knitr session. It maybe the HTML or a link to the pdf output for example.
}
\description{
depending on the context return the correct format for a document.
The basic output here is to use HTML as an output if possible and convert it to an image or a PDF that can then
be included into a latex document for example.
}
\concept{output}
