% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-utils.R
\name{hux_set_footer}
\alias{hux_set_footer}
\title{Set a huxtable caption as a first row}
\usage{
hux_set_footer(hux, footer)
}
\arguments{
\item{hux}{a huxtable}

\item{footer}{footer text}
}
\value{
a huxtable with first row caption
}
\description{
Keeps the same formatting as the rest of the table
}
