% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-utils.R
\name{hux_sprintf}
\alias{hux_sprintf}
\title{A sprintf alternative that handles NA values gracefully (ish)}
\usage{
hux_sprintf(fmt, ..., na.text = "—")
}
\arguments{
\item{fmt}{sprintf format string}

\item{...}{sprintf inputs}

\item{na.text}{an string to replave NA values with.}
}
\value{
a strgin
}
\description{
A sprintf alternative that handles NA values gracefully (ish)
}
