% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-utils.R
\name{hux_to_ggplot}
\alias{hux_to_ggplot}
\title{Convert a huxtable to a ggplot object}
\usage{
hux_to_ggplot(hux, width = 5.9)
}
\arguments{
\item{hux}{the huxtable}

\item{width}{the desired ggplot width}
}
\value{
a ggplot object of the right width
}
\description{
Useful if you need to include a formatted table in a figure with a plot
}
