% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{intersecting_group_by}
\alias{intersecting_group_by}
\title{Create a new data frame including duplicate rows where the rows fulfil a potentially overlapping set of conditions}
\usage{
intersecting_group_by(.data, ..., .colname = "group")
}
\arguments{
\item{.data}{a data frame}

\item{...}{a set of predicates specified like case_whens syntax, such as mpg < 5 ~ "gas guzzlers"}

\item{.colname}{the name of the new group}
}
\value{
a new dataframe containing the overlapping groups.
}
\description{
Create a new data frame including duplicate rows where the rows fulfil a potentially overlapping set of conditions
}
\examples{
iris \%>\% group_by(Species) \%>\% intersecting_group_by(
  Sepal.Length > mean(Sepal.Length) ~ "Long",
  Sepal.Width > mean(Sepal.Width) ~ "Wide"
)
}
