% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-utils.R
\name{is_installed}
\alias{is_installed}
\title{check if a package is installed}
\usage{
is_installed(packageName)
}
\arguments{
\item{packageName}{the name of the package}
}
\value{
boolean value
}
\description{
check if a package is installed
}
