% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{knit_print.rendered_plot}
\alias{knit_print.rendered_plot}
\title{Knit a rendered_plot object}
\usage{
\method{knit_print}{rendered_plot}(x, options, ...)
}
\arguments{
\item{x}{the rendered_plot}

\item{options}{the chunk options}

\item{...}{not used}
}
\value{
nothing - used for side effects
}
\description{
Knit a rendered_plot object
}
