% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{knit_print.rendered_table}
\alias{knit_print.rendered_table}
\title{Knit a rendered_table object}
\usage{
\method{knit_print}{rendered_table}(x, ...)
}
\arguments{
\item{x}{the rendered_table}

\item{...}{not used}
}
\value{
nothing - used for side effects
}
\description{
Knit a rendered_table object
}
\concept{output}
