% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-output-utils.R
\name{knit_versioned}
\alias{knit_versioned}
\title{Knit to a versioned file in a sub-directory of the project}
\usage{
knit_versioned(
  ext,
  directory = NULL,
  datedFile = !datedSubdirectory,
  datedSubdirectory = FALSE
)
}
\arguments{
\item{directory}{the root of the output - can be an absolute path or a relative path interpreted as relative to the root of the project.}

\item{datedFile}{do you want the filename to have the date appended?}

\item{datedSubdirectory}{do you want the files to be placed in a dated subdirectory?}

\item{exts}{the extension of file to target e.g. "pdf", "html" (TODO: multiple)}
}
\value{
nothing. called for side effects
}
\description{
Knit to a versioned file in a sub-directory of the project
}
