% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{loadfonts}
\alias{loadfonts}
\title{Version of loadfonts from extrafonts that uses first font file it finds rather than failing if multiple are found}
\usage{
loadfonts(device = "pdf", quiet = FALSE)
}
\arguments{
\item{device}{The output device. Can be "pdf" (the default), "postscript", or "win".}

\item{quiet}{If FALSE, print a status message as each font is registered. If TRUE, don't print.}
}
\value{

}
\description{
Version of loadfonts from extrafonts that uses first font file it finds rather than failing if multiple are found
}
