% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-struct-list.R
\name{map2_struct}
\alias{map2_struct}
\title{Map over two inputs}
\usage{
map2_struct(.x, .y, .f, ..., .progress = FALSE)
}
\arguments{
\item{.x}{a `struct_list`}

\item{.y}{a list or vector the same length as `.x`}

\item{.f}{a function to apply to each `.x`, `.y` pair}

\item{...}{additional parameters to  `.f` but anonymous function preferred}

\item{.progress}{display a progress bar (logical or string name)}
}
\value{
a `struct_list`
}
\description{
These functions are variants of \code{\link[purrr:map]{map()}} that iterate over two arguments at
a time.
}
\concept{structures}
