% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-struct-list.R
\name{Math.struct_list}
\alias{Math.struct_list}
\title{Maths operations}
\usage{
\method{Math}{struct_list}(x, ...)
}
\description{
abs, sign, sqrt, floor, ceiling, trunc, round, signif
exp, log, expm1, log1p,
cos, sin, tan, cospi, sinpi, tanpi, acos, asin, atan, cosh, sinh, tanh, acosh, asinh, atanh
lgamma, gamma, digamma, trigamma
cumsum, cumprod, cummax, cummin
}
\examples{
# example code

abs.test = function(x,...) {
  x$a = abs(x$a)
  return(x)
}

x = struct(a=-1,b=2,c=1:3,.class="test")
y = struct(a=4,b=5,c=1:3,.class="test")

abs.test(x[[1]])

# TODO: further testing
# abs(rep(c(x,y),5))

}
\concept{structures}
\keyword{internal}
