% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-utils.R
\name{non_cran}
\alias{non_cran}
\title{Make sure github packages are installed.}
\usage{
non_cran(name, github, force = FALSE, subdir = "", ...)
}
\arguments{
\item{name}{the name of the package}

\item{github}{something like "github-repo/project-name"}

\item{force}{will only update a loaded package if TRUE (defaults to FALSE)}

\item{subdir}{if the package is in a subdirectory of the github repo}

\item{...}{passed to devtools::install_github}
}
\value{
nothing
}
\description{
Use a locally checked out version if available.
}
\examples{
non_cran("patchwork", "thomasp85/patchwork")
}
