% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-struct-list.R
\name{Ops.struct_list}
\alias{Ops.struct_list}
\title{Binary operations}
\usage{
\method{Ops}{struct_list}(e1, e2)
}
\description{
Binary operations
}
\examples{

as.double.test = function(x) {
  return(x$a)
}

`+.test` = function(e1,e2) {
  e1$a = e1$a + as.numeric(e2)
  return(e1)
}

x = struct(a=1,b=2,c=1:3,.class="test")

# TODO: needs debugging
# x+3
# x[[1]]+3
# rep(x,5)+1:5
# x[[1]]+x[[1]]

}
\concept{structures}
\keyword{internal}
