% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-output-utils.R
\name{outputter}
\alias{outputter}
\title{Generate a versioned file name in a subdirectory.}
\usage{
outputter(
  directory = here::here("output"),
  datedFile = !datedSubdirectory,
  datedSubdirectory = FALSE
)
}
\arguments{
\item{directory}{the root of the output}

\item{datedFile}{do you want the filename to have the date appended?}

\item{datedSubdirectory}{do you want the files to be placed in a dated subdirectory?}
}
\value{
a function that takes a filename and boolean delete parameter.
When called with a filename component this function will return the absolute path of a file which is
versioned with date. If the file exists and delete=TRUE it is deleted
(allowing for libraries that refuse to overwrite existing files)
}
\description{
This function generates a function that resolves a file path fragment to a
specific file location, accounting for a versioning strategy involving
the current date. The defaults create a naming strategy that places an
file in the "output" sub-directory of the current project with a filename suffix including
the date.
}
\examples{
library(tidyverse)
out = outputter("~/output",datedSubdirectory=TRUE)
out("file.png")
}
