% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{pbeta2}
\alias{pbeta2}
\title{The Beta Distribution}
\usage{
pbeta2(q, prob, kappa, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles}

\item{prob}{the mean probability (vectorised)}

\item{kappa}{a coefficient of variation. where 0 is no variability and 1 is
maximally variability (vectorised)}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}
otherwise \code{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
\code{dbeta} gives the density, \code{pbeta} the distribution
  function, \code{qbeta} the quantile function, and \code{rbeta}
  generates random deviates.

  Invalid arguments will result in return value \code{NaN}, with a warning.

  The length of the result is determined by \code{n} for
  \code{rbeta}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.
}
\description{
Density, distribution function, quantile function and random
  generation for the Beta distribution with parameters \code{shape1} and
  \code{shape2} (and optional non-centrality parameter \code{ncp}).
}
\examples{
pbeta2(c(0.25,0.5,0.75), 0.5, 0.25)
}
\seealso{
\code{\link[stats:Beta]{stats::pbeta()}}
}
\concept{distributions}
