% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-ggplot-utils.R
\name{percent_trans}
\alias{percent_trans}
\title{percentage scale}
\usage{
percent_trans()
}
\value{
A scales object
}
\description{
display a 0-1 scale as 0-100\%
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\concept{ggplot}
