% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{plnorm2}
\alias{plnorm2}
\title{The Log Normal Distribution}
\source{
\code{dlnorm} is calculated from the definition (in \sQuote{Details}).
  \code{[pqr]lnorm} are based on the relationship to the normal.

  Consequently, they model a single point mass at \code{exp(meanlog)}
  for the boundary case \code{sdlog = 0}.
}
\usage{
plnorm2(q, mean = 1, sd = sqrt(exp(1) - 1), lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles}

\item{mean}{the mean value on the true scale (vectorised)}

\item{sd}{the standard deviation on the true scale (vectorised)}

\item{lower.tail}{logical; if TRUE (default), probabilities are
    \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
\code{dlnorm} gives the density,
  \code{plnorm} gives the distribution function,
  \code{qlnorm} gives the quantile function, and
  \code{rlnorm} generates random deviates.

  The length of the result is determined by \code{n} for
  \code{rlnorm}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.
}
\description{
Density, distribution function, quantile function and random
  generation for the log normal distribution whose logarithm has mean
  equal to \code{meanlog} and standard deviation equal to \code{sdlog}.
}
\details{
The log normal distribution has density
  \deqn{
    f(x) = \frac{1}{\sqrt{2\pi}\sigma x} e^{-(\log(x) - \mu)^2/2 \sigma^2}%
  }{f(x) = 1/(\\sqrt(2 \\pi) \\sigma x) e^-((log x - \\mu)^2 / (2 \\sigma^2))}
  where \eqn{\mu} and \eqn{\sigma} are the mean and standard
  deviation of the logarithm.
  The mean is \eqn{E(X) = exp(\mu + 1/2 \sigma^2)},
  the median is \eqn{med(X) = exp(\mu)}, and the variance
  \eqn{Var(X) = exp(2\mu + \sigma^2)(exp(\sigma^2) - 1)}{Var(X) = exp(2*\\mu + \\sigma^2)*(exp(\\sigma^2) - 1)}
  and hence the coefficient of variation is
  \eqn{\sqrt{exp(\sigma^2) - 1}}{sqrt(exp(\\sigma^2) - 1)} which is
  approximately \eqn{\sigma} when that is small (e.g., \eqn{\sigma < 1/2}).
}
\note{
The cumulative hazard \eqn{H(t) = - \log(1 - F(t))}{H(t) = - log(1 - F(t))}
  is \code{-plnorm(t, r, lower = FALSE, log = TRUE)}.
}
\examples{
plnorm2(seq(0,4,0.25), 2, 1)
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.

  Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995)
  \emph{Continuous Univariate Distributions}, volume 1, chapter 14.
  Wiley, New York.
}
\seealso{
\link[stats]{Distributions} for other standard distributions, including
  \code{\link[stats]{dnorm}} for the normal distribution.
}
\concept{distributions}
