% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\docType{data}
\name{presidential}
\alias{presidential}
\alias{std_size}
\title{Standard image and paper sizes}
\format{
A list
}
\usage{
data(presidential)
}
\description{
The width and height of images to fit scientific publication standards
}
\keyword{datasets}
