% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{print.rendered_plot}
\alias{print.rendered_plot}
\title{Print a rendered_plot object}
\usage{
\method{print}{rendered_plot}(x, ...)
}
\arguments{
\item{x}{the rendered_plot}

\item{...}{not used}
}
\value{
nothing - used for side effects
}
\description{
Print a rendered_plot object
}
\concept{output}
