% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-utils.R
\name{print.rendered_table}
\alias{print.rendered_table}
\title{Print a rendered_table object}
\usage{
\method{print}{rendered_table}(x, ...)
}
\arguments{
\item{x}{the rendered_table}

\item{...}{not used}
}
\value{
nothing - used for side effects
}
\description{
Print a rendered_table object
}
