% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{qgamma2}
\alias{qgamma2}
\title{The Gamma Distribution}
\usage{
qgamma2(
  p,
  mean,
  sd = sqrt(mean),
  lower.tail = TRUE,
  log.p = FALSE,
  convex = TRUE
)
}
\arguments{
\item{p}{vector of probabilities}

\item{mean}{the mean value on the true scale (vectorised)}

\item{sd}{the standard deviation on the true scale (vectorised)}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}
otherwise \code{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{convex}{Show a warning if the distribution selected is not a convex
function}
}
\value{
\code{dgamma} gives the density,
  \code{pgamma} gives the distribution function,
  \code{qgamma} gives the quantile function, and
  \code{rgamma} generates random deviates.

  Invalid arguments will result in return value \code{NaN}, with a warning.

  The length of the result is determined by \code{n} for
  \code{rgamma}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.
}
\description{
Density, distribution function, quantile function and random
  generation for the Gamma distribution with parameters \code{shape} and
  \code{scale}.
}
\examples{
qgamma2(c(0.25,0.5,0.75), 2, 1)
}
\seealso{
\code{\link[stats:GammaDist]{stats::qgamma()}}
}
\concept{distributions}
