% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rbern}
\alias{rbern}
\title{A random Bernoulli sample as a logical value}
\usage{
rbern(n, prob)
}
\arguments{
\item{n}{number of observations}

\item{prob}{the mean probability (vectorised)}
}
\value{
a vector of logical values of size \code{n}
}
\description{
A random Bernoulli sample as a logical value
}
\examples{
table(rbern(100, 0.25))
}
\concept{distributions}
