% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rbeta2}
\alias{rbeta2}
\title{The Beta Distribution}
\usage{
rbeta2(n, prob, kappa)
}
\arguments{
\item{n}{number of observations}

\item{prob}{the mean probability (vectorised)}

\item{kappa}{a coefficient of variation. where 0 is no variability and 1 is
maximally variability (vectorised)}
}
\value{
\code{dbeta} gives the density, \code{pbeta} the distribution
  function, \code{qbeta} the quantile function, and \code{rbeta}
  generates random deviates.

  Invalid arguments will result in return value \code{NaN}, with a warning.

  The length of the result is determined by \code{n} for
  \code{rbeta}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.
}
\description{
Density, distribution function, quantile function and random
  generation for the Beta distribution with parameters \code{shape1} and
  \code{shape2} (and optional non-centrality parameter \code{ncp}).
}
\examples{
rbeta2(3, c(0.1,0.5,0.9),0.1)
}
\seealso{
\code{\link[stats:Beta]{stats::rbeta()}}
}
\concept{distributions}
