% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rcategorical}
\alias{rcategorical}
\title{Sampling from the multinomial equivalent of the Bernoulli distribution}
\usage{
rcategorical(n, prob, factor = FALSE)
}
\arguments{
\item{n}{a sample size}

\item{prob}{a (optionally named) vector of probabilities that will be
normalised to sum to 1}

\item{factor}{if not FALSE then factor levels will either be taken from the
names of \code{prob} first, or if this is a character vector from this.}
}
\value{
a vector of random class labels of length \code{n}. Labels come from names
of \code{prob} or from a character vector in \code{factor}.
}
\description{
Sampling from the multinomial equivalent of the Bernoulli distribution
}
\examples{

prob = c("one"=0.1,"two"=0.2,"seven"=0.7)
table(rcategorical(1000,prob))
rcategorical(10,prob,factor=TRUE)
rcategorical(10,rep(1,26),factor=letters)

}
\concept{distributions}
