% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{register_web_fonts}
\alias{register_web_fonts}
\title{Find fonts in webfont providers}
\usage{
register_web_fonts(fonts, services = names(webfont_provider), ...)
}
\arguments{
\item{fonts}{a set of fonts as a character vector}

\item{services}{one of `r paste0("'",names(webfont_provider),"'",collapse=", ")`}

\item{...}{not used}
}
\value{
a list containing the following:
* webfonts - a list of `svglite::font_face` directives for use in embedding
into svg files using the `svglite::svglite(web_fonts=...)` option.
* matched - a dataframe of fonts found in the services
* unmatched - a vector of fonts not matched online.
}
\description{
Caches, registers and constructs webfont CSS font face directives.
}
\examples{
tmp = register_web_fonts("Kings")
tmp$webfonts
systemfonts::registry_fonts()
}
