% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{reparam-dist}
\alias{reparam-dist}
\title{Re-parametrised distributions}
\arguments{
\item{n}{number of observations}

\item{x}{vector of quantiles}

\item{q}{vector of quantiles}

\item{p}{vector of probabilities}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are `P[X<=x]`
otherwise `P[X>x]`.}

\item{convex}{Show a warning if the distribution selected is not a convex
function}

\item{prob}{the mean probability (vectorised)}

\item{kappa}{a coefficient of variation. where 0 is no variability and 1 is
maximally variability (vectorised)}

\item{mean}{the mean value on the true scale (vectorised)}

\item{sd}{the standard deviation on the true scale (vectorised)}
}
\description{
Re-parametrised distributions
}
\concept{distributions}
