% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rgamma2}
\alias{rgamma2}
\title{The Gamma Distribution}
\usage{
rgamma2(n, mean, sd = sqrt(mean), convex = TRUE)
}
\arguments{
\item{n}{number of observations}

\item{mean}{the mean value on the true scale (vectorised)}

\item{sd}{the standard deviation on the true scale (vectorised)}

\item{convex}{Show a warning if the distribution selected is not a convex
function}
}
\value{
\code{dgamma} gives the density,
  \code{pgamma} gives the distribution function,
  \code{qgamma} gives the quantile function, and
  \code{rgamma} generates random deviates.

  Invalid arguments will result in return value \code{NaN}, with a warning.

  The length of the result is determined by \code{n} for
  \code{rgamma}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.
}
\description{
Density, distribution function, quantile function and random
  generation for the Gamma distribution with parameters \code{shape} and
  \code{scale}.
}
\examples{
rgamma2(10, 2, 1)
}
\seealso{
\code{\link[stats:GammaDist]{stats::rgamma()}}
}
\concept{distributions}
