% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-ggplot-utils.R
\name{scale_fill_consistent}
\alias{scale_fill_consistent}
\title{Use a colour scale from one plot in another}
\usage{
scale_fill_consistent(
  plot,
  original_aesthetic = "fill",
  target_aesthetic = "fill",
  ...
)
}
\arguments{
\item{plot}{a ggplot with a colour scale to clone}

\item{original_aesthetic}{the original aesthetic we are cloning (fill or color)}

\item{target_aesthetic}{the aesthetic in the new plot we want to match.}
}
\description{
Multiple factor levels with custom labels are difficult to sync between
graphs if some levels are missing in the second plot. This copies the
palette from one plot to another. It is sometimes the case that we want to reuse
the fill from one plot as the color for another.
}
\concept{ggplot}
