% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-ggplot-utils.R
\name{scale_y_logit}
\alias{scale_y_logit}
\title{A logit y scale}
\usage{
scale_y_logit(..., n = 5, sf = 2)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}

\item{n}{the number of major breaks}

\item{sf}{significant figures}
}
\value{
a ggplot scale
}
\description{
A logit y scale
}
\concept{ggplot}
