% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-ggplot-utils.R
\name{scale_y_percent}
\alias{scale_y_percent}
\title{A percentage y scale}
\usage{
scale_y_percent(..., sf = 2)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}

\item{sf}{significant figures}
}
\value{
a ggplot scale
}
\description{
A percentage y scale
}
\concept{ggplot}
