% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-struct-list.R
\name{struct_flatten}
\alias{struct_flatten}
\title{Ensure \code{struct_list} is a flat list of \code{structs}}
\usage{
struct_flatten(x)
}
\arguments{
\item{x}{a potentially nested list of \code{struct_lists}.}
}
\value{
a flat \code{struct_list} of \code{structs}
}
\description{
Unlike \code{purrr::list_flatten} this is recursive for one reason
which is that a \code{struct_list} must only contain \code{structs}.
}
\concept{structures}
