% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{summarise_with_totals}
\alias{summarise_with_totals}
\title{Summarise and include a total row, or a row including the summary for the whole group, into a factor list.}
\usage{
summarise_with_totals(
  .data,
  ...,
  .groups = NULL,
  .total = "Total",
  .total_first = FALSE
)
}
\arguments{
\item{.data}{a dataframe}

\item{...}{the summarisation specification}

\item{.groups}{what to do with the grouping after summarisation (same as dplyr::summarise)}

\item{.total}{name of the total row which will be added into a factor list.}

\item{.total_first}{should the total be before or after the groups}
}
\value{
a dataframe with the additional
}
\description{
Summarise and include a total row, or a row including the summary for the whole group, into a factor list.
}
\examples{
diamonds \%>\% group_by(color,cut) \%>\%
   summarise_with_totals(mpg = sprintf("\%1.1f \u00B1 \%1.1f", mean(price), sd(price)), .total = "Overall")
}
