% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-struct-list.R
\name{Summary.struct_list}
\alias{Summary.struct_list}
\title{Summary operations}
\usage{
\method{Summary}{struct_list}(..., na.rm = FALSE)
}
\description{
all, any, sum, prod, min, max, range
}
\examples{

sum.test = function(..., na.rm=TRUE) {
  # must use rlang::list2
  l = rlang::list2(...)
  return(purrr::reduce(l, `+.test`))
}

map_struct(1:10, ~ struct(a=.x,b=2,c=1:3,.class="test"))

min.test = function(..., na.rm=TRUE) {
  l = rlang::list2(...)

}

# class(sum(z$b))
# sum(z$b, z$b)

}
\concept{structures}
\keyword{internal}
