% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-deprecated-dependency-utils.R
\name{unstable}
\alias{unstable}
\title{Get unstable version of package}
\usage{
unstable(pkg = "ggrrr", org = "terminological")
}
\arguments{
\item{pkg}{the package to load the unstable version of}

\item{org}{the github organisation}
}
\value{
nothing
}
\description{
reboot a library reloading from a local development copy if it exists locally alternatively get
the most up to date github package.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# ggrrr::unstable()
}
