% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{bind_rows_with_factors}
\alias{bind_rows_with_factors}
\title{Bind_rows works until there are factors with a set of different levels.}
\usage{
bind_rows_with_factors(...)
}
\arguments{
\item{...}{a list of dataframes}
}
\value{
the union of those dataframes. Factor levels are combined with levels
}
\description{
Bind_rows works until there are factors with a set of different levels.
}
\examples{
bind_rows_with_factors(iris,diamonds)
}
