% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{check_font}
\alias{check_font}
\title{Ensures a font is available.}
\usage{
check_font(family)
}
\arguments{
\item{family}{a font family name or names}
}
\value{
the font family name if it can be located or an error otherwise
}
\description{
This checks to see if a font exists. If missing it will try and install from google fonts.
If this is not possible it will throw an error.
}
\examples{
check_font(c("Roboto","Arial","Kings","EB Garamond"))
}
