% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-utils.R
\name{cran}
\alias{cran}
\title{Make sure cran packages are installed}
\usage{
cran(cran_deps)
}
\arguments{
\item{cran_deps}{a vector of package names}
}
\value{
nothing
}
\description{
Make sure cran packages are installed
}
\examples{
cran("tidyverse")
}
