% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.gg_save_as}
\alias{.gg_save_as}
\title{Save a plot to multiple formats}
\usage{
.gg_save_as(
  plot,
  filename = tempfile(),
  size = std_size$half,
  maxWidth = size$width,
  maxHeight = size$height,
  aspectRatio = maxWidth/maxHeight,
  formats = getOption("ggrrr.formats", default = c("svg", "png", "pdf")),
  webfontFinder = ~return(list()),
  ...
)
}
\arguments{
\item{plot}{a ggplot}

\item{filename}{base of target filename (excluding extension).}

\item{size}{a standard size see `std_size`}

\item{maxWidth}{maximum width in inches}

\item{maxHeight}{maximum height in inches}

\item{aspectRatio}{defaults to maxWidth/maxHeight}

\item{formats}{some of svg, png, pdf, Rdata, eps, ...}

\item{webfontFinder}{a function that takes a plot and returns a properly
formatted css specification for webfonts in the plot. This is for internal
use and does not need to be changed.}

\item{...}{
  Named arguments passed on to \code{\link[svglite:svglite]{svglite::svglite}}\describe{
    \item{\code{pointsize}}{Default point size.}
\item{\code{standalone}}{Produce a standalone svg file? If \code{FALSE}, omits
xml header and default namespace.}
\item{\code{id}}{A character vector of ids to assign to the generated SVG's. If
creating more SVG files than supplied ids the exceeding SVG's will not have
an id tag and a warning will be thrown.}
\item{\code{fix_text_size}}{Should the width of strings be fixed so that it doesn't
change between svg renderers depending on their font rendering? Defaults to
\code{TRUE}. If \code{TRUE} each string will have the \code{textLength} CSS property set
to the width calculated by systemfonts and
\code{lengthAdjust='spacingAndGlyphs'}. Setting this to \code{FALSE} can be
beneficial for heavy post-processing that may change content or style of
strings, but may lead to inconsistencies between strings and graphic
elements that depend on the dimensions of the string (e.g. label borders
and background).}
\item{\code{scaling}}{A scaling factor to apply to the rendered line width and text
size. Useful for getting the right sizing at the dimension that you
need.}
\item{\code{always_valid}}{Should the svgfile be a valid svg file while it is being
written to? Setting this to \code{TRUE} will incur a considerable performance
hit (>50\% additional rendering time) so this should only be set to \code{TRUE}
if the file is being parsed while it is still being written to.}
\item{\code{file}}{Identical to \code{filename}. Provided for backward compatibility.}
}}
}
\value{
the output is an sensible default object that can be displayed given
  the context it is called in, for example if knitting an RMarkdown document
  a link to the png file for embedding, if latex a link to the pdf file.
}
\description{
Saves a ggplot object to disk at a set physical size. Allows specific maximum dimensions
with an optional target aspect ratio to fit into specific configurations for publication.
e.g. a half page plot or a third of a 2 column page. Allows output in pdf for journal
publication or png for inclusion in documents, and makes sure that the outputs are
near identical.
}
\details{
For maximum cross platform reproducibility we are using the combination of
`systemfonts` for font management, `svglite` to render the canonical output
`rsvg` to convert that to pdf, and `ragg` for bitmap formats.
In some situations `rsvg` fails in which case we fall back to rendering in a
headless chrome instance. This rather complicated pipeline ensures modern
webfont support, and editable SVG or PDF.
}
\concept{output}
\keyword{internal}
\keyword{plot}
