% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggplot-utils.R
\name{.gg_scale_consistent}
\alias{.gg_scale_consistent}
\title{Use a colour scale from one plot in another}
\usage{
.gg_scale_consistent(
  plot,
  original_aesthetic = c("fill", "color"),
  target_aesthetic = original_aesthetic,
  ...
)
}
\arguments{
\item{plot}{a ggplot with a colour scale to clone}

\item{original_aesthetic}{the original aesthetic we are cloning (fill or color)}

\item{target_aesthetic}{the aesthetic in the new plot we want to match.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual}}
  \describe{
    \item{\code{values}}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with \code{breaks} if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that don't
match will be given \code{na.value}.}
    \item{\code{aesthetics}}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}
    \item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output
}}
    \item{\code{na.value}}{The aesthetic value to use for missing (\code{NA}) values}
  }}
}
\description{
Multiple factor levels with custom labels are difficult to sync between
graphs if some levels are missing in the second plot. This copies the
palette from one plot to another. It is sometimes the case that we want to reuse
the fill from one plot as the color for another.
}
\concept{ggplot}
\keyword{internal}
