% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.here}
\alias{.here}
\title{Drop in replacement for `here` (`here` pkg)}
\usage{
.here(..., projRoot = .locate_project())
}
\arguments{
\item{...}{the relative path within the project}

\item{projRoot}{the project root - defaults to `.locate_project()`}
}
\value{
a path
}
\description{
Drop in replacement for `here` (`here` pkg)
}
\examples{
try(.here("vignettes"))
}
\concept{output}
\keyword{internal}
