% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.pmixgamma}
\alias{.pmixgamma}
\title{The cumulative density function of a mixture of gamma distributions}
\usage{
.pmixgamma(q, shapes, rates, weights = rep(1, length(shapes)), na.rm = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{shapes}{a vector of gamma distribution shapes}

\item{rates}{a vector of gamma distribution rates}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions which have NA for shape or rate}
}
\value{
the pdf of the mixture distribution.
}
\description{
The cumulative density function of a mixture of gamma distributions
}
\examples{
try({
  .pmixgamma(q=c(2,20), shapes=c(10,13,14), rates=c(1,1,1), weights=c(2,2,3))
})
}
\concept{distributions}
\keyword{internal}
