% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.pmixlnorm}
\alias{.pmixlnorm}
\title{The cumulative density function of a mixture of log normal distributions}
\usage{
.pmixlnorm(q, meanlogs, sdlogs, weights = 1, na.rm = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{meanlogs}{a vector of normal distribution means}

\item{sdlogs}{a vector of normal distribution sds}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions which have NA for mean or sd}
}
\value{
the pdf of the mixture distribution.
}
\description{
The cumulative density function of a mixture of log normal distributions
}
\concept{distributions}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
.pmixlnorm(q=c(2,20), meanlogs=c(1.0,1.3,1.4), sdlogs=c(1,1,2), weights=c(2,2,3))
}\if{html}{\out{</div>}}
}

