% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.print_html_with_chrome}
\alias{.print_html_with_chrome}
\title{Print a HTML fragment using chrome}
\usage{
.print_html_with_chrome(
  htmlFragment,
  pdfFile = tempfile(fileext = ".pdf"),
  css = list(),
  chromeBinary = getOption("ggrrr.chrome", default = .find_chrome()),
  maxWidth = NULL,
  maxHeight = NULL
)
}
\arguments{
\item{htmlFragment}{an HTML fragment}

\item{pdfFile}{the pdf file}

\item{css}{a list of valid css style specifications (not urls to stylesheets)}

\item{chromeBinary}{the chrome browser (or chromium) path}

\item{maxWidth}{the maximum width of the output in inches}

\item{maxHeight}{the maximum height of the output in inches}
}
\value{
the pdf path
}
\description{
Print a HTML fragment using chrome
}
\concept{output}
\keyword{internal}
