% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-file-output-utils.R
\name{.print_svg_with_chrome}
\alias{.print_svg_with_chrome}
\title{Print an SVG file to pdf using chrome}
\usage{
.print_svg_with_chrome(
  svgFile,
  pdfFile = tempfile(fileext = ".pdf"),
  chromeBinary = getOption("ggrrr.chrome", default = .find_chrome()),
  maxWidth = NULL,
  maxHeight = NULL
)
}
\arguments{
\item{svgFile}{the input svg file}

\item{pdfFile}{the output pdf file}

\item{chromeBinary}{the location of chrome or chromium binary}

\item{maxWidth}{the maximum width of the output in inches}

\item{maxHeight}{the maximum height of the output in inches}
}
\value{
the pdf file path
}
\description{
Create a the same size as the SVG file using chrome or chromium
}
\concept{output}
\keyword{internal}
