% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.qmixbeta}
\alias{.qmixbeta}
\title{A quantile function for a mixture of gamma distributions}
\usage{
.qmixbeta(
  p,
  alphas,
  betas,
  weights = 1,
  na.rm = FALSE,
  method = c("exact", "samples", "moments"),
  ...
)
}
\arguments{
\item{p}{vector of probabilities.}

\item{alphas}{a vector of gamma distribution shapes}

\item{betas}{a vector of gamma distribution rates}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions with NA values for mean or sd}

\item{method}{one of `exact` (solve with uniroot), `samples` (random resampling), `moments` (Cornish Fisher approximation)}

\item{...}{passed on to internal function, `seed=XXX` to fix random seed}
}
\value{
the value of the `p`th quantile
}
\description{
A quantile function for a mixture of gamma distributions
}
\examples{
try({
  .qmixbeta(p=c(0.025,0.5,0.975), alphas=c(10,13,14), betas=c(1,1,2))
  .qmixbeta(p=c(0.025,0.5,0.975), alphas=c(10,13,14), betas=c(1,1,2), method="moments")
})
}
\concept{distributions}
\keyword{internal}
