% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.qmixgamma}
\alias{.qmixgamma}
\title{A quantile function for a mixture of gamma distributions}
\usage{
.qmixgamma(
  p,
  shapes,
  rates,
  weights = 1,
  na.rm = FALSE,
  method = c("exact", "samples", "moments"),
  ...
)
}
\arguments{
\item{p}{vector of probabilities.}

\item{shapes}{a vector of gamma distribution shapes}

\item{rates}{a vector of gamma distribution rates}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions with NA values for mean or sd}

\item{method}{one of \code{exact} (solve with uniroot), \code{samples} (random resampling), \code{moments} (Cornish Fisher approximation)}

\item{...}{passed on to internal function, \code{seed=XXX} to fix random seed}
}
\value{
the value of the \code{p}th quantile
}
\description{
A quantile function for a mixture of gamma distributions
}
\concept{distributions}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
.qmixgamma(p=c(0.025,0.5,0.975), shapes=c(10,13,14), rates=c(1,1,2), method="moments")
.qmixgamma(p=c(0.025,0.5,0.975), shapes=c(10,13,14), rates=c(1,1,2), method="exact")
}\if{html}{\out{</div>}}
}

