% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.qmixgamma}
\alias{.qmixgamma}
\title{A quantile function for a mixture of gamma distributions}
\usage{
.qmixgamma(
  p,
  shapes,
  rates,
  weights = 1,
  na.rm = FALSE,
  method = c("exact", "samples", "moments"),
  ...
)
}
\arguments{
\item{p}{vector of probabilities.}

\item{shapes}{a vector of gamma distribution shapes}

\item{rates}{a vector of gamma distribution rates}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions with NA values for mean or sd}

\item{method}{one of `exact` (solve with uniroot), `samples` (random resampling), `moments` (Cornish Fisher approximation)}

\item{...}{passed on to internal function, `seed=XXX` to fix random seed}
}
\value{
the value of the `p`th quantile
}
\description{
A quantile function for a mixture of gamma distributions
}
\concept{distributions}
\keyword{internal}
