% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.qmixlnorm}
\alias{.qmixlnorm}
\title{A quantile function for a mixture of log normal distributions}
\usage{
.qmixlnorm(
  p,
  meanlogs,
  sdlogs,
  weights = rep(1, length(meanlogs)),
  na.rm = FALSE,
  approx = TRUE,
  ...
)
}
\arguments{
\item{p}{vector of probabilities.}

\item{meanlogs}{a vector of log normal distribution means}

\item{sdlogs}{a vector of log normal distribution sds}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions with NA values for mean or sd}

\item{approx}{use Cornish Fisher moment based approximation if available}

\item{...}{passed on to internal function, `samples=TRUE` to force random sampling}
}
\value{
the value of the `p`th quantile
}
\description{
A quantile function for a mixture of log normal distributions
}
\examples{
try({
  .qmixlnorm(p=c(0.025,0.5,0.975), meanlogs=c(1,1.3,1.4), sdlogs=c(0.1,0.1,0.2))
  .qmixlnorm(p=c(0.025,0.5,0.975), meanlogs=c(1,1.3,1.4), sdlogs=c(0.1,0.1,0.2), approx=FALSE)
  .qmixlnorm(p=c(0.025,0.5,0.975), meanlogs=c(1,1.3,1.4), sdlogs=c(1,1,2), approx=FALSE)
})
}
\concept{distributions}
\keyword{internal}
