% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-mixture-distributions.R
\name{.qmixnorm}
\alias{.qmixnorm}
\title{A quantile function for a mixture of normal distributions}
\usage{
.qmixnorm(
  p,
  means,
  sds,
  weights = 1,
  na.rm = FALSE,
  method = c("exact", "samples", "moments"),
  ...
)
}
\arguments{
\item{p}{vector of probabilities.}

\item{means}{a vector of normal distribution means}

\item{sds}{a vector of normal distribution sds}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions with NA values for mean or sd}

\item{...}{passed on to internal function, `samples=TRUE` to force random sampling}

\item{approx}{use Cornish Fisher moment based approximation if available}
}
\value{
the value of the `p`th quantile
}
\description{
A quantile function for a mixture of normal distributions
}
\examples{
try({
.qmixnorm(p=c(0.025,0.5,0.975), means=c(10,13,14), sds=c(1,1,2))
})
}
\concept{distributions}
\keyword{internal}
