% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-ggplot-utils.R
\name{drawDetails.watermark}
\alias{drawDetails.watermark}
\title{Internal function for drawing watermark on ggplots}
\usage{
\method{drawDetails}{watermark}(x, recording)
}
\arguments{
\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\value{
a grid object
}
\description{
Internal function for drawing watermark on ggplots
}
\concept{ggplot}
