% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-utils.R
\name{fit_col_widths}
\alias{fit_col_widths}
\title{Estimate column content widths}
\usage{
fit_col_widths(table)
}
\arguments{
\item{table}{a table to get column content widths for.}
}
\value{
a vector of column widths
}
\description{
Widths are based on dataframe or huxtable content ignoring rowspans and
potential for wrapping.
}
\examples{
library(tidyverse)
iris \%>\% fit_col_widths()
}
