% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaprogramming-utils.R
\name{fix_unqualified_functions}
\alias{fix_unqualified_functions}
\title{Fix errors introduced in package creation by forgetting to qualify namespaces.}
\usage{
fix_unqualified_functions(
  rDirectories = c(here::here("R"), here::here("vignettes")),
  description = here::here("DESCRIPTION"),
  dry_run = TRUE
)
}
\arguments{
\item{rDirectories}{the locations of the R code to fix}

\item{description}{the location of the description files}

\item{dry_run}{by default this function will not actually do anything unless this is set to FALSE.}
}
\value{

}
\description{
Fix errors introduced in package creation by forgetting to qualify namespaces.
}
