% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{fonts_available}
\alias{fonts_available}
\title{Which fonts are available on this system.}
\usage{
fonts_available(family = NULL)
}
\arguments{
\item{family}{a font family name or names}
}
\value{
the font family name if it can be located or an error otherwise
}
\description{
Which fonts are available on this system.
}
\examples{
fonts_available(c("Arial","sdfdsfsd"))
}
