% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{fonts_available}
\alias{fonts_available}
\title{Check a font exists on the system, or list all fonts available}
\usage{
fonts_available(family = NULL)
}
\arguments{
\item{family}{an optional list of fonts to check for}
}
\value{
a list of available fonts
}
\description{
Given a font family this checks it si locally available in both the
`showtext` and `systemfonts` subsystems
}
\examples{
fonts_available(c("Arial","sdfdsfsd"))
}
