% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-tidy-utils.R
\name{get_value_sets}
\alias{get_value_sets}
\title{Get a value set list of a dataframe}
\usage{
get_value_sets(df)
}
\arguments{
\item{df}{a dataframe to examine}
}
\value{
a list of lists with the column name and the factor levels as list, as a \verb{checked list}.
}
\description{
This function examines a dataframe and returns a list of the columns with sub-lists as all the options for factors.
This provides programmatic access (and autocomplete) to the values available in a dataframe, and throws and early
error if we try and access data by a variable that does not exist.
}
\concept{tidydata}
