% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-ggplot-utils.R
\name{gg_pedantic}
\alias{gg_pedantic}
\title{An opinionated set of defaults for plots}
\usage{
gg_pedantic(
  lineSize = 0.25,
  fontSize = 8,
  font = "Roboto",
  size = lineSize * 2,
  ...
)
}
\arguments{
\item{lineSize}{the default line and shape size in ggplot units}

\item{fontSize}{the base font size}

\item{font}{the default font name.}

\item{size}{the size of points (the default size aesthetic)}
}
\value{
nothing
}
\description{
This is a set of styles with a focus on making plots compact, and minimally
fussy, and ensuring fonts are consistent between axes and labels. It sets
default sizes for line widths and point sizes. It also
switched the default png renderer in knitr to `ragg::ragg_png` to allow for
modern font support.
}
\concept{ggplot}
