% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-ggplot-utils.R
\name{gg_simple_table}
\alias{gg_simple_table}
\title{A simple table as a ggplot patchwork object, no customisation allowed}
\usage{
gg_simple_table(df, pts = 8, font = "sans", unwrapped = FALSE)
}
\arguments{
\item{df}{the dataframe with the table data. Column names will become headings}

\item{pts}{text size in points}

\item{font}{the font family}

\item{unwrapped}{\itemize{
\item set this to TRUE if you want to add to a patchwork and use patchwork::wrap_plots(p,list(table))
}}
}
\value{
A gtable object (i.e. a grob) optionally wrapped as a patchwork plot.
}
\description{
A simple table as a ggplot patchwork object, no customisation allowed
}
\examples{
if (FALSE) {
  gg_simple_table(tibble::tibble(x=c(1,2,3),y=c(5,4,3)),pts=10)
}
}
\concept{ggplot}
\keyword{graph}
\keyword{layout}
