% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{gg_tiny_theme}
\alias{gg_tiny_theme}
\title{A space saving ggplot theme}
\usage{
gg_tiny_theme(baseSize = 8, font = "Roboto")
}
\arguments{
\item{baseSize}{the size of the base font.}

\item{font}{the font family name}
}
\value{
a ggplot theme
}
\description{
A ggplot theme with minimal fluff and with the defaults set small.
}
\examples{
if (interactive()) {
  ggplot2::ggplot(ggplot2::diamonds,
    ggplot2::aes(x=carat,y=price,color=color))+
    ggplot2::geom_point()+
    gg_tiny_theme()
}
}
