% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glyphmap}
\alias{glyphmap}
\title{The glyphmap dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item agl (character) - the agl: adobe glyph label
\item unicode (character) - the unicode code
}

4281 rows and 2 columns
}
\usage{
data(glyphmap)
}
\description{
The adobe glyph name to UTF cross map.
}
\details{
from \verb{https://raw.githubusercontent.com/adobe-type-tools/agl-aglfn/refs/heads/master/glyphlist.txt}
}
\concept{fonts}
\keyword{datasets}
