% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{google_font}
\alias{google_font}
\title{Downloads caches and installs a font family from google}
\usage{
google_font(family, ...)
}
\arguments{
\item{family}{the google font family (e.g. Roboto) as found here: https://fonts.google.com/ (case sensitive)}

\item{...}{not used}
}
\value{
nothing
}
\description{
Downloads caches and installs a font family from google
}
