% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utils.R
\name{knit_versioned}
\alias{knit_versioned}
\title{Knit to a versioned file in a sub-directory of the project}
\usage{
knit_versioned(
  directory = NULL,
  ...,
  datedFile = !datedSubdirectory,
  datedSubdirectory = FALSE
)
}
\arguments{
\item{directory}{the root of the output - can be an absolute path or a relative path interpreted as relative to the root of the project.}

\item{...}{ignored}

\item{datedFile}{do you want the filename to have the date appended (defaults TRUE)?}

\item{datedSubdirectory}{do you want the files to be placed in a dated subdirectory  (defaults FALSE)?}
}
\value{
nothing. called for side effects
}
\description{
used in a knitr preamble to direct the output to a subdirectory of the project
---
title: "Analysis 1"
output: html_document
knit: ggrrr::knit_versioned("output/analysis-1")
---
}
\details{
This can only work when deployed as a library and hence no standalone version
of it exists, because the fully qualified packagename has to be used.
}
