% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-output-utils.R
\name{outputter}
\alias{outputter}
\title{Create a function to generate versioned file name in a subdirectory.}
\usage{
outputter(
  directory = here::here("output"),
  datedFile = !datedSubdirectory,
  datedSubdirectory = FALSE
)
}
\arguments{
\item{directory}{the root of the output}

\item{datedFile}{do you want the filename to have the date appended}

\item{datedSubdirectory}{do you want the files to be placed in a dated subdirectory}
}
\value{
a function that takes a filename and boolean delete parameter. The filename will be the relative path within the configured directory. If the file exists and delete=TRUE it is deleted (for libraries that refuse to overwrite existing files)
}
\description{
Create a function to generate versioned file name in a subdirectory.
}
\examples{
out = outputter("~/output",datedSubdirectory=TRUE)
out("file.png")
}
