% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-struct-list.R
\name{pmap_struct}
\alias{pmap_struct}
\title{Map over multiple input simultaneously (in "parallel")}
\usage{
pmap_struct(.l, .f, ..., .progress = FALSE)
}
\arguments{
\item{.l}{A \code{struct_list}.}

\item{.f}{a function to apply to each \code{.l} item}

\item{...}{additional parameters to  \code{.f} but anonymous function preferred}

\item{.progress}{display a progress bar (logical or string name)}
}
\value{
a \code{struct_list}
}
\description{
These functions are variants of \code{\link[purrr:map]{map()}} that iterate over multiple arguments
simultaneously. They are parallel in the sense that each input is processed
in parallel with the others, not in the sense of multicore computing, i.e.
they share the same notion of "parallel" as \code{\link[base:Extremes]{base::pmax()}} and \code{\link[base:Extremes]{base::pmin()}}.
}
\concept{structures}
