% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-api-ggplot-utils.R
\name{scale_y_logit}
\alias{scale_y_logit}
\title{A logit y scale}
\usage{
scale_y_logit(..., n = 5, sf = 2)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}

\item{n}{the number of major breaks}

\item{sf}{significant figures}
}
\value{
a ggplot scale
}
\description{
A logit y scale
}
\examples{
tibble::tibble(pvalue = c(0.001, 0.05, 0.1), fold_change = 1:3) \%>\%
 ggplot2::ggplot(ggplot2::aes(fold_change , pvalue)) +
 ggplot2::geom_point() +
 .gg_scale_y_logit(n=8)
}
\concept{ggplot}
