% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-struct-list.R
\name{struct_flatten}
\alias{struct_flatten}
\title{Ensure `struct_list` is a flat list of `structs`}
\usage{
struct_flatten(x)
}
\arguments{
\item{x}{a potentially nested list of `struct_lists`.}
}
\value{
a flat `struct_list` of `structs`
}
\description{
Unlike `purrr::list_flatten` this is recursive for one reason
which is that a `struct_list` must only contain `structs`.
}
\concept{structures}
