% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-utils.R
\name{unstable}
\alias{unstable}
\title{Get unstable version of package}
\usage{
unstable(pkg = "ggrrr", org = "terminological")
}
\arguments{
\item{pkg}{the package to load the unstable version of}

\item{org}{the github organisation}
}
\value{
nothing
}
\description{
reboot a library reloading from a local development copy if it exists locally alternatively get
the most up to date github package.
}
\examples{
\donttest{
library(tidyverse)
ggrrr::unstable()
}
}
