% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-utils.R
\name{unstable}
\alias{unstable}
\title{reboot this library reloading from dev.}
\usage{
unstable(pkg = "ggrrr")
}
\value{
nothing
}
\description{
reboot this library reloading from dev.
}
\examples{
ggrrr::unstable()
}
